+++ packages/core/src/config/setupLogger.ts
@@ -1,11 +1,10 @@
-import { consoleTransport, Logger } from '@abimongo/logger'; 
+import { consoleTransport, Logger } from '@abimongo/logger';
 
 
 import {
   bufferedTransporter,
   elasticTransport,
 } from '../utils';
-
 
 Logger.initialize({
   formatOptions: {