+++ packages/core/src/index.ts
@@ -1,7 +1,7 @@
 /**
  * @author Emmanuel Nodolomwanyi - Abimongo Team
  * @package - @abimongo/core
- * @version 1.0.0
+ * @version 1.1.3
  */
 
 /**
@@ -9,7 +9,7 @@
  * This library provides core functionalities for Abimongo, ORM/ODM solution
  * for MongoDB in Node.js and Browser environments.
  * @module AbimongoCore
- * @version 1.0.0
+ * @version 1.1.3
  */