+++ packages/create/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @abimongo/create
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @abimongo/logger@2.0.0
+
 ## 1.0.3
 
 ### Patch Changes